// src/api/video-v1/model/mediaFile.ts
/**
 * Drtv Videos Service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeliveryTypeEnum } from './deliveryTypeEnum';
import { ContainerFormatEnum } from './containerFormatEnum';
import { Subtitles } from './subtitles';
import { ResolutionEnum } from './resolutionEnum';
import { TimeCodes } from './timeCodes';


export interface MediaFile { 
    accessService: string;
    name: string;
    deliveryType: DeliveryTypeEnum;
    url: string;
    drm: string;
    complianceProfile: string | null;
    containerFormat?: ContainerFormatEnum;
    format: string;
    resolution: ResolutionEnum;
    width: number;
    height: number;
    channels?: number | null;
    language: string;
    subtitles?: Array<Subtitles> | null;
    isStreamLive?: boolean | null;
    timeCodes?: Array<TimeCodes> | null;
}